+++ point.h
@@ -4,7 +4,7 @@
 
 class Point{ 
 public:
-	int** adjacentMatrix;
+	float** adjacentMatrix;
 	int pointNum; //the number of generated pointset
 	set< pair<int,int> > pointset;
 
@@ -14,10 +14,10 @@ public:
 	void generatePoint(unsigned int H, unsigned int W, unsigned int N);
 	set< pair<int,int> > getPointset();
 	void setPointset();
-	int ** getAdjacentMatrix();
+	float ** getAdjacentMatrix();
 	int getPointNum();
 	void printPointset();
 
 private:
-	int getEuclideanDistance(int x1, int y1, int x2, int y2);
+	float getEuclideanDistance(int x1, int y1, int x2, int y2);
 };
